requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.4
supabase>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
slowapi>=0.1.9
